motion: Change units of ec_rate from us to ms.
To ease finer calculation of ec rate change units from
ms to us.

BRANCH=smaug
BUG=b:24367625
TEST=compile

+++ include/motion_sense.h
@@ -50,14 +50,14 @@ struct motion_data_t {
 	 * data rate the sensor will measure, in mHz: 0 suspended.
 	 * MSB is used to know if we are rounding up.
 	 * */
-	unsigned odr;
+	unsigned int odr;
 	/* delay between collection by EC, in ms.
 	 * For non FIFO sensor, should be need 1e6/odr to
 	 * collect events.
 	 * For sensor with FIFO, can be much longer.
 	 * 0: no collection.
 	 */
-	unsigned short ec_rate;
+	unsigned int ec_rate;
 };
 
 struct motion_sensor_t {
